requirements.txt
apscheduler>=3.10
requests>=2.31
timezonefinder>=6.2
pytz>=2023.3
geopy>=2.4
twilio>=8.0
python-dotenv>=1.0
pytest>=7.0
pytest-cov>=4.0
